Exercise 12:
Comunicated child to parent. Feature: if you click on add, the console
will prompt the data of that product.

The product component now creates an event which outputs the product
data to the products component. On the products component html we added
the event and handled with a function the console.log that we wanted.
The function was implemented at products component ts.