Exercise 13:
Added a new shared service shopping-cart to handle the addition of a new
product to the cart. We defined three observables which will be updating the
total price, the number of products and each of the products added to the cart.

We called this function inside products component when we click on add product.
Finally, on the header component we showed the total price and the number of products.