Exercise 22:
Created a new module for a thank you page after we checkout.
Added a function at shopping cart shared service to reset the cart.
Added the route to the checkout routing module.
Added the logic at checkout component to call the reset cart function
and to navigate to the thank you page.